# Conflicts:
#	HW2/src/BattleManager.cpp

+++ HW2/src/hw2.cpp
@@ -3,12 +3,7 @@
 
 #include "BattleManager.h"
 #include "BattleShipGameFromFile.h"
-#include <iostream>
-#include <cstdio>
-
-#include <ctype.h>
 #include <stdio.h>
-#include <stdlib.h>
 #include "ConsolePrinter.h"
 #include "Utility.h"
 
@@ -27,7 +22,7 @@ int main(int argc, char* argv[])
 	if (result.first) {
 		if (result.second != "") {
 			try {
-				delay = atoi(result.second.c_str()) * 1000;
+				delay = (int)(stod(result.second.c_str()) * 1000);
 			}
 			catch (exception) {
